lambdas/authorizer/requirements.txt
python-jose[cryptography]>=3.3
